pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "path-builder-turnback"
version = "0.1.0"
description = "TurnBack public release: 36kroutes, Path Builder, route generation, and reverse-route prompting."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Hongyi Luo" }]
dependencies = [
    "geopandas>=1.0.1",
    "matplotlib>=3.9.2",
    "networkx>=3.3",
    "numpy>=2.1.0",
    "osmnx>=1.9.4",
    "pandas>=2.2.2",
    "pyproj>=3.6.1",
    "requests>=2.32.3",
    "scipy>=1.13.1",
    "shapely>=2.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.2",
]
llm = [
    "google-generativeai>=0.8.4",
    "openai>=1.54.0",
]

[project.scripts]
path-builder = "path_builder.cli:main"

[project.urls]
Repository = "https://github.com/bghjmn32/EMNLP2025_Turnback"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-include = ["*.ipynb"]
